refactor: :recycle: credentials lookup
\
