+++ app/Controllers/Http/ImageController.js
@@ -4,19 +4,11 @@ const Env = use("Env");
 const AWS = require("aws-sdk");
 const { v4: uuid4 } = require("uuid");
 
-if (
-  process.env.AWS_ACCESS_KEY_ID &&
-  process.env.AWS_SECRET_ACCESS_KEY &&
-  process.env.AWS_REGION
-) {
-  // do nothing, the aws-sdk will read those environment variables
-} else {
-  AWS.config.credentials = new AWS.SharedIniFileCredentials({
-    profile: Env.get("AWS_PROFILE_NAME", "default"),
-  });
-}
-
 class ImageController {
+  static get inject() {
+    require("../../Common/Credentials");
+  }
+
   // uploads image to S3, returns image url
   async uploadImage({ request, response }) {
     const { region } = request.all();
